requirements/common.txt
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.7
beautifulsoup4==4.7.1
billiard==3.6.3.0
celery==4.4.2
certifi==2019.6.16
chardet==3.0.4
cssutils==1.0.2
Django==2.2.12
django-autoslug==1.9.4
django-bootstrap3==12.0.3
django-choices==1.7.1
django-filter==2.2.0
djangorestframework==3.11.0
fake-useragent==0.1.11
idna==2.8
importlib-metadata==1.6.0
kombu==4.6.8
lxml==4.4.1
pkg-resources==0.0.0
psycopg2-binary==2.8.3
python-decouple==3.1
pytz==2019.1
requests==2.22.0
requests-file==1.4.3
six==1.13.0
soupsieve==1.9.1
sqlparse==0.3.0
tldextract==2.2.2
urllib3==1.25.3
vine==1.3.0
zipp==3.1.0
